requirements.txt
altgraph==0.17.3
amadeus==12.0.0
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
autopep8==2.0.1
azure-common==1.1.28
azure-core==1.32.0
azure-eventgrid==4.21.0
azure-functions==1.21.3
azure-identity==1.19.0
azure-mgmt-core==1.5.0
azure-mgmt-resource==23.2.0
azure-mgmt-storage==22.0.0
azure-mgmt-synapse==2.0.0
azure-storage-blob==12.24.0
azure-synapse-spark==0.7.0
beautifulsoup4==4.12.2
blinker==1.9.0
cached-property==1.5.2
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
convertdate==2.4.0
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.5
django-cors-headers==4.2.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
docutils==0.20.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
exchangelib==4.9.0
Faker==19.6.2
fire==0.4.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.39.4
fpdf2==2.7.4
furl==2.1.3
graphviz==0.13.2
h11==0.14.0
hijri-converter==2.3.1
holidays==0.21.13
htmldocx==0.0.6
hubspot-api-client==4.0.6
idna==3.4
importlib-metadata==4.13.0
intelhex==2.3.0
isodate==0.6.1
itsdangerous==2.2.0
java-access-bridge-wrapper==0.10.0
Jinja2==3.1.5
jsonpath-ng==1.5.3
jsonschema==4.17.3
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
korean-lunar-calendar==0.3.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==3.0.2
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
mss==6.1.0
mysql-connector-python==8.0.32
netsuitesdk==1.24.0
notifiers==1.3.3
numpy==1.26.4
O365==2.0.26
oauthlib==3.2.2
openpyxl==3.1.2
orderedmultidict==1.0.1
outcome==1.2.0
packaging==23.1
pandas==2.2.1
pdfminer.six==20221105
pefile==2023.2.7
pendulum==2.1.2
Pillow==9.5.0
platformdirs==3.5.1
ply==3.11
portalocker==2.10.1
psutil==5.9.5
pyarrow==19.0.0
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.15.1
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
PyJWT==2.7.0
PyMeeus==0.5.12
pynput-robocorp-fork==5.0.0
pyodbc==5.2.0
pyotp==2.8.0
pypdf==3.8.1
pyperclip==1.8.2
pypiwin32==223
pyrsistent==0.19.3
PySocks==1.7.1
pyspnego==0.9.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
pywin32==303
pywin32-ctypes==0.2.0
pywinauto==0.6.8
PyYAML==6.0
requests==2.30.0
requests-file==1.5.1
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
robotframework==5.0.1
robotframework-pythonlibcore==4.1.2
robotframework-requests==0.9.4
robotframework-sapguilibrary==1.1
robotframework-seleniumlibrary==6.1.0
robotframework-seleniumtestability==2.1.0
rpaframework==23.0.0
rpaframework-core==11.0.1
rpaframework-pdf==7.1.5
rpaframework-windows==7.3.2
selenium==4.9.1
simple-salesforce==1.12.3
six==1.16.0
smartsheet-python-sdk==3.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
sqlparse==0.4.3
stringcase==1.2.0
tenacity==8.2.2
termcolor==2.3.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==3.10.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2022.7
tzlocal==4.3
uiautomation==2.0.18
urllib3==2.0.2
validators==0.20.0
webdriver-manager==3.8.6
Werkzeug==3.1.3
wrapt==1.15.0
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.2.1
zipp==3.15.0
